pyproject.toml
[project]
name = "kpi-etl"
version = "0.1.0"
description = "ETL pipeline for ДП «Ліси України» KPI dashboard — parses Excel into canonical SQLite schema"
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "sshtunnel>=0.4",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-pyyaml>=6.0",
    "types-openpyxl>=3.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RET"]
ignore = ["E501"]  # line-length handled by formatter

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_reexport = true

[[tool.mypy.overrides]]
module = ["openpyxl.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-xvs --tb=short"
